Add (package) objects to seenClasses.
This enables us to extract methods calls that look like

    PackageObjectName.this.Wrapper(arg)

such as implicit conversions derived from implicit classes.

Eg. with

object TopLevelObject {

    implicit class BoolOps(x: Boolean) {
      def ==>(y: Boolean) = !x || y
    }

    val foo = true ==> false

}

scalac will produce the following tree

object TopLevelObject {

    implicit class BoolOps(x: Boolean) {
      def ==>(y: Boolean) = !x || y
    }

    implicit def BoolOps(x: Boolean): TopLevelObject.BoolOps = new BoolOps(x)

    val foo = TopLevelObject.this.BoolOps(true) ==> false

}

Note the `TopLevelObject.this` path prefix in the definition of `foo`.

To extract those calls, we need to keep track of whether we have already seen a (package) object
with a similar symbol, which is what this commit does.

There might be a cleaner way but I haven't found it yet.