Extract implicit classes as case classes.
This allows us to extract the following code

    object BoolStuff {

      implicit class BoolOps(x: Boolean) {
        def ==>(y: Boolean) = !x || y
      }

      val test = true ==> false

    }

as

    object BoolStuff {

      case class BoolOps(x: Boolean) {
          def ==>(y: Boolean) = !x || y
      }

      val test = BoolOps(true) ==> false

    }

The implicit conversion

generated by scalac is lost but it shouldn't be too hard to extract it as well.

For the record, scalac doesn't actually instanciate the implicit class,
but instead defines and applies an implicit conversion wich eventually instanciate it.
Here's what scalac actually generates:

    def BoolOps(x: Boolean): BoolOps = new BoolOps(x)

    val test = BoolStuff.this.BoolOps(x) ==> false

If needed, it shouldn't be too hard to extract the conversion and its application as-is.